Fix up Dynflow action to update task status in database
Previously, we were relying on the front end calling job_status and job_result from the Task Execution page to update the database with results from the smart proxy. This means if the user navigated away during task execution and never came back, the status would never be accurate in the database or Task History page. This fixes the problems that were in the PollTaskStatus action so that it can correctly poll and update status in the background, and the endpoints used by the front end just pull data from the database rather than reaching out to the smart proxy themselves and updating when called.