Fix for build
The build container needs the rubygem-foreman-tasks package installed
first. This also changes things to use our own copy of the plugin spec
erb file rather than the default from foreman-packaging in case we need
to tweak it in the future, and adds foreman-tasks as an explicit runtime
dependency on the package.